All apis have been tested
apis tested with hardcoded apis.
changes made to each api and routes to allow integration with token authorization.

+++ backend/controllers/inventoryController.js
@@ -7,12 +7,10 @@ const Admin = require('../models/adminModel')
 //vendor views their product stock
 const viewInventory = async(req, res) => {
     try{
-        /*const userID = req.user.userID
+        const userID = req.user.userID
 
         const admin = await Admin.findById(userID)
-        const vendor = admin.vendorId*/
-
-        const vendor = new ObjectId('6410a78a09a73ac03605b9e6')
+        const vendor = admin.vendorId
 
         const inventoryView = await Inventory.find({ vendorId: vendor }).populate('productId', 'name');
 
@@ -38,12 +36,10 @@ const viewInventory = async(req, res) => {
 //vendor adds new product
 const addProduct = async(req, res) => {
     try{
-        /*const userID = req.user.userID
+        const userID = req.user.userID
 
         const admin = await Admin.findById(userID)
-        const vendor = admin.vendorId*/
-
-        const vendor = new ObjectId('6410a78a09a73ac03605b9e6')
+        const vendor = admin.vendorId
 
         const newProduct = new Product({
             name : req.body.name,
@@ -86,28 +82,36 @@ const addProduct = async(req, res) => {
 //vendor deletes their product
 const deleteProduct = async(req, res) => {
     try{
-        /*const userID = req.user.userID
+        const userID = req.user.userID
 
         const admin = await Admin.findById(userID)
-        const vendor = admin.vendorId*/
+        const vendor = admin.vendorId
 
-        const vendor = new ObjectId('6410a78a09a73ac03605b9e6')
 
         const Stock = await Inventory.findOne({productId: req.body.productId})
 
-        try{
-            const removeStock = await Inventory.deleteOne({productId: req.body.productId})
-            const removeProduct = await Product.deleteOne({productId: removeStock.productId})
-            res.json({
-                success: true,
-                error: "Product Deleted!"
-            })
-        } catch(err) {
+        if(Stock.vendorId === vendor) {
+            try{
+                const removeStock = await Inventory.deleteOne({productId: req.body.productId})
+                const removeProduct = await Product.deleteOne({productId: removeStock.productId})
+                res.json({
+                    success: true,
+                    error: "Product Deleted!"
+                })
+            } catch(err) {
+                res.json({
+                    success: false,
+                    error: err.message
+                })
+            }
+        }
+        else{
             res.json({
                 success: false,
-                error: err.message
+                data: "Unauthorized to perform this action"
             })
         }
+        
         
     } catch(err) {
         res.json({
@@ -120,14 +124,7 @@ const deleteProduct = async(req, res) => {
 //vendor views a specific product
 const viewProduct = async(req, res) => {
     try {
-        /*const userID = req.user.userID
-
-        const admin = await Admin.findById(userID)
-        const vendor = admin.vendorId*/
-
-        const vendor = new ObjectId('6410a78a09a73ac03605b9e6')
-
-        const product = await Product.find({_id : req.body.productId} && {vendorID : vendor});
+        const product = await Product.find({_id : req.body.productId});
 
         res.status(200).json({
             succuss: true,
@@ -144,29 +141,41 @@ const viewProduct = async(req, res) => {
 //vendor updates product and its stock
 const updateProduct = async(req, res) => {
     try{
-        /*const userID = req.user.userID
+        const userID = req.user.userID
 
         const admin = await Admin.findById(userID)
-        const vendor = admin.vendorId*/
-        const vendor = new ObjectId('6410a78a09a73ac03605b9e6')
+        const vendor = admin.vendorId
 
         const product = await Product.findById(req.body.productId)
-
-        product.name = req.body.name;
-        product.description = req.body.description;
-        product.image = req.body.image;
-        const savedProduct = await product.save();
-
         const productStock = await Inventory.find({productId: product._id})
-        if(productStock.length>0){
-            const stock = productStock[0];
-            stock.discount = req.body.discount;
-            stock.price = req.body.price;
-            stock.quantity = req.body.quantity;
 
-            const savedStock = await stock.save();
+        if(vendor === productStock.vendorId){
+            product.name = req.body.name;
+            product.description = req.body.description;
+            product.image = req.body.image;
+            const savedProduct = await product.save();
+
+            
+            if(productStock.length>0){
+                const stock = productStock[0];
+                stock.discount = req.body.discount;
+                stock.price = req.body.price;
+                stock.quantity = req.body.quantity;
+
+                const savedStock = await stock.save();
+            }
+            res.json({
+                success: true,
+                data: "Product updated Successfully!"
+            });
         }
-        res.json("Product updated Successfully!");
+        else {
+            res.json({
+                success: false,
+                data: "Unauthorized to perform this action"
+            })
+        }
+        
     } catch(err) {
         res.json({
             success: false,
